Modify SpriteManager and GameControl
So the game can no add an arrow when you press the 'S' key, but the
arrow is not being drawn, can someone better than me at programming take
a look in the SpriteManager class at the SpawnArrow() method and then in
turn at the Draw(GameTime gameTime) method because i honestly have no
idea why it's not drawing the arrow onto the screen.
I'm gonna move on to working with Cave, Map, and GUI to display a room
when you enter the game, along with the separate doors. i won't be here
tomorrow morning, but i'll try to get something done during the 20
minutes that i'm in the waiting room.